New API - Add test cases - break out in-dialog session tests - add tests for re-INVITE BYE race conditions
