+++ test/spec/api/session.spec.ts
@@ -1152,6 +1152,63 @@ describe("API Session", () => {
         });
       });
 
+      describe("Bob accept(), 200 send has no SDP - Invalid 200 Ok", () => {
+        beforeEach(async () => {
+          resetSpies();
+          return invitation.accept({ test: "acceptWithoutDescription" })
+            .then(() => bob.transport.waitReceived());
+        });
+
+        it("her ua should send ACK", () => {
+          const spy = alice.transportSendSpy;
+          expect(spy).toHaveBeenCalledTimes(2);
+          expect(spy.calls.argsFor(0)).toEqual(SIP_ACK);
+          expect(spy.calls.argsFor(1)).toEqual(SIP_BYE);
+        });
+
+        it("her ua should receive 200", () => {
+          const spy = alice.transportReceiveSpy;
+          expect(spy).toHaveBeenCalledTimes(2);
+          expect(spy.calls.mostRecent().args).toEqual(SIP_200);
+          expect(spy.calls.mostRecent().args).toEqual(SIP_200);
+        });
+
+        it("her request delegate should onAccept", () => {
+          const spy = inviterRequestDelegateMock;
+          expect(spy.onAccept).toHaveBeenCalledTimes(1);
+          expect(spy.onProgress).toHaveBeenCalledTimes(0);
+          expect(spy.onRedirect).toHaveBeenCalledTimes(0);
+          expect(spy.onReject).toHaveBeenCalledTimes(0);
+          expect(spy.onTrying).toHaveBeenCalledTimes(0);
+        });
+
+        it("her session state should transition 'terminated'", () => {
+          const spy = inviterStateSpy;
+          expect(spy).toHaveBeenCalledTimes(1);
+          expect(spy.calls.argsFor(0)).toEqual([SessionState.Terminated]);
+        });
+
+        it("his session state should transition 'establishing', 'established', 'terminated'", () => {
+          const spy = invitationStateSpy;
+          expect(spy).toHaveBeenCalledTimes(3);
+          expect(spy.calls.argsFor(0)).toEqual([SessionState.Establishing]);
+          expect(spy.calls.argsFor(1)).toEqual([SessionState.Established]);
+          expect(spy.calls.argsFor(2)).toEqual([SessionState.Terminated]);
+        });
+
+        it("her dialog should be 'terminated' and 'closed'", () => {
+          const session = inviter.dialog;
+          expect(session && session.sessionState).toBe(SessionDialogState.Terminated);
+          expect(session && session.signalingState).toBe(SignalingState.Closed);
+        });
+
+        it("his dialog should be 'terminated' and 'closed'", () => {
+          const session = invitation.dialog;
+          expect(session && session.sessionState).toBe(SessionDialogState.Terminated);
+          expect(session && session.signalingState).toBe(SignalingState.Closed);
+        });
+      });
+
       describe("Bob accept(), 200 send fails - Transport Error", () => {
         beforeEach(async () => {
           if (!(bob.userAgent.transport instanceof TransportFake)) {
@@ -2015,594 +2072,6 @@ describe("API Session", () => {
       });
 
       inviteSuiteFork(true);
-    });
-  });
-
-  function reinviteAccepted(withoutSdp: boolean): void {
-    beforeEach(async () => {
-      resetSpies();
-      invitation.delegate = undefined;
-      return inviter.invite({ withoutSdp })
-        .then(() => alice.transport.waitSent()); // ACK
-    });
-
-    it("her ua should send INVITE, ACK", () => {
-      const spy = alice.transportSendSpy;
-      expect(spy).toHaveBeenCalledTimes(2);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-      expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-    });
-
-    it("her ua should receive 200", () => {
-      const spy = alice.transportReceiveSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_200);
-    });
-
-    it("her signaling should be stable", () => {
-      if (!inviter.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(inviter.dialog.signalingState).toEqual(SignalingState.Stable);
-    });
-
-    it("his signaling should be stable", () => {
-      if (!invitation.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(invitation.dialog.signalingState).toEqual(SignalingState.Stable);
-    });
-  }
-
-  function reinviteAcceptedWithoutDescriptionFailure(withoutSdp: boolean): void {
-    beforeEach(async () => {
-      resetSpies();
-      invitation.delegate = { onReinviteTest: () => "acceptWithoutDescription" };
-      const session: Session = inviter;
-      return session.invite({ withoutSdp })
-        .then(() => alice.transport.waitSent()); // ACK
-    });
-
-    it("her ua should send INVITE, ACK, BYE", () => {
-      const spy = alice.transportSendSpy;
-      expect(spy).toHaveBeenCalledTimes(3);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-      expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-      expect(spy.calls.argsFor(2)).toEqual(SIP_BYE);
-    });
-
-    it("her ua should receive 200", () => {
-      const spy = alice.transportReceiveSpy;
-      expect(spy).toHaveBeenCalledTimes(2);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_200);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_200);
-    });
-
-    it("her signaling should be closed", () => {
-      if (!inviter.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(inviter.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("his signaling should be closed", () => {
-      if (!invitation.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(invitation.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("her session state should transition 'terminated'", () => {
-      const spy = inviterStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("his session state should transition 'terminated'", () => {
-      const spy = invitationStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("her session should be failed", () => {
-      expect(inviter.isFailed).toBe(true);
-    });
-  }
-
-  function reinviteAcceptedOfferAnswerFailure(withoutSdp: boolean): void {
-    beforeEach(async () => {
-      resetSpies();
-      invitation.delegate = undefined;
-      const session: Session = inviter;
-      return session.invite({ withoutSdp })
-        .then(() => {
-          const sessionDescriptionHandler = session.sessionDescriptionHandler;
-          if (!sessionDescriptionHandler) {
-            throw new Error("Session description handler undefined");
-          }
-          const sdh = sessionDescriptionHandler as jasmine.SpyObj<Required<SessionDescriptionHandler>>; // assumes a spy
-          sdh.getDescription.and.callFake(() => Promise.reject(new Error("Failed to get description.")));
-          sdh.setDescription.and.callFake(() => Promise.reject(new Error("Failed to set description.")));
-        })
-        .then(() => alice.transport.waitSent()); // ACK
-    });
-
-    it("her ua should send INVITE, ACK, BYE", () => {
-      const spy = alice.transportSendSpy;
-      expect(spy).toHaveBeenCalledTimes(3);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-      expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-      expect(spy.calls.argsFor(2)).toEqual(SIP_BYE);
-    });
-
-    it("her ua should receive 200", () => {
-      const spy = alice.transportReceiveSpy;
-      expect(spy).toHaveBeenCalledTimes(2);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_200);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_200);
-    });
-
-    it("her signaling should be closed", () => {
-      if (!inviter.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(inviter.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("his signaling should be closed", () => {
-      if (!invitation.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(invitation.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("her session state should transition 'terminated'", () => {
-      const spy = inviterStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("his session state should transition 'terminated'", () => {
-      const spy = invitationStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("her session should be failed", () => {
-      expect(inviter.isFailed).toBe(true);
-    });
-  }
-
-  function reinviteRejected(withoutSdp: boolean): void {
-    beforeEach(async () => {
-      resetSpies();
-      invitation.delegate = { onReinviteTest: () => "reject488" };
-      const session: Session = inviter;
-      return session.invite({ withoutSdp })
-        .then(() => alice.transport.waitSent()); // ACK
-    });
-
-    it("her ua should send INVITE, ACK", () => {
-      const spy = alice.transportSendSpy;
-      expect(spy).toHaveBeenCalledTimes(2);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-      expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-    });
-
-    it("her ua should receive 488", () => {
-      const spy = alice.transportReceiveSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_488);
-    });
-
-    it("her signaling should be stable", () => {
-      if (!inviter.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(inviter.dialog.signalingState).toEqual(SignalingState.Stable);
-    });
-
-    it("his signaling should be stable", () => {
-      if (!invitation.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(invitation.dialog.signalingState).toEqual(SignalingState.Stable);
-    });
-  }
-
-  function reinviteRejectedRollbackFailure(withoutSdp: boolean): void {
-    beforeEach(async () => {
-      resetSpies();
-      invitation.delegate = { onReinviteTest: () => "reject488" };
-      const session: Session = inviter;
-      return session.invite({ withoutSdp: false }) // Note that rollback on reject this only happens INVITE with SDP
-        .then(() => {
-          const sessionDescriptionHandler = session.sessionDescriptionHandler;
-          if (!sessionDescriptionHandler) {
-            throw new Error("Session description handler undefined");
-          }
-          const sdh = sessionDescriptionHandler as jasmine.SpyObj<Required<SessionDescriptionHandler>>; // assumes a spy
-          sdh.rollbackDescription.and.callFake(() => Promise.reject(new Error("Failed to rollback description.")));
-        })
-        .then(() => alice.transport.waitSent());  // ACK
-    });
-
-    it("her ua should send INVITE, ACK, BYE", () => {
-      const spy = alice.transportSendSpy;
-      expect(spy).toHaveBeenCalledTimes(3);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-      expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-      expect(spy.calls.argsFor(2)).toEqual(SIP_BYE);
-    });
-
-    it("her ua should receive 488", () => {
-      const spy = alice.transportReceiveSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)).toEqual(SIP_488);
-    });
-
-    it("her signaling should be closed", () => {
-      if (!inviter.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(inviter.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("his signaling should be closed", () => {
-      if (!invitation.dialog) {
-        fail("Session dialog undefined");
-        return;
-      }
-      expect(invitation.dialog.signalingState).toEqual(SignalingState.Closed);
-    });
-
-    it("her session state should transition 'terminated'", () => {
-      const spy = inviterStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("his session state should transition 'terminated'", () => {
-      const spy = invitationStateSpy;
-      expect(spy).toHaveBeenCalledTimes(1);
-      expect(spy.calls.argsFor(0)[0]).toEqual(SessionState.Terminated);
-    });
-
-    it("her session should be failed", () => {
-      expect(inviter.isFailed).toBe(true);
-    });
-  }
-
-  function reinviteSuite(withoutSdp: boolean): void {
-    describe("Alice invite()", () => {
-      beforeEach(() => {
-        resetSpies();
-        return inviter.invite()
-          .then(() => bob.transport.waitSent());
-      });
-
-      describe("Bob accept()", () => {
-        beforeEach(() => {
-          resetSpies();
-          invitation.delegate = undefined;
-          return invitation.accept()
-            .then(() => alice.transport.waitSent()); // ACK
-        });
-
-        it("her state should be `established`", () => {
-          expect(inviter.state).toBe(SessionState.Established);
-        });
-
-        it("his state should be `established`", () => {
-          expect(invitation.state).toBe(SessionState.Established);
-        });
-
-        describe("Alice invite() accepted", () => {
-          reinviteAccepted(withoutSdp);
-
-          describe("Alice invite() accepted", () => {
-            reinviteAccepted(withoutSdp);
-          });
-        });
-
-        describe("Alice invite() accepted without description failure", () => {
-          reinviteAcceptedWithoutDescriptionFailure(withoutSdp);
-        });
-
-        describe("Alice invite() accepted offer/answer failure", () => {
-          reinviteAcceptedOfferAnswerFailure(withoutSdp);
-        });
-
-        describe("Alice invite() rejected", () => {
-          reinviteRejected(withoutSdp);
-
-          describe("Alice invite() accepted", () => {
-            reinviteAccepted(withoutSdp);
-          });
-        });
-
-        describe("Alice invite() rejected rollback failure", () => {
-          reinviteRejectedRollbackFailure(withoutSdp);
-        });
-      });
-    });
-  }
-
-  describe("In Dialog...", () => {
-    beforeEach(async () => {
-      target = bob.uri;
-      bob.userAgent.delegate = {
-        onInvite: (session) => {
-          invitation = session;
-          invitationEmitSpy = makeEventEmitterEmitSpy(invitation, bob.userAgent.getLogger("Bob"));
-          invitationStateSpy = makeEmitterSpy(invitation.stateChange, bob.userAgent.getLogger("Bob"));
-        }
-      };
-      inviter = new Inviter(alice.userAgent, target);
-      inviterEmitSpy = makeEventEmitterEmitSpy(inviter, alice.userAgent.getLogger("Alice"));
-      inviterStateSpy = makeEmitterSpy(inviter.stateChange, alice.userAgent.getLogger("Alice"));
-      await soon();
-    });
-
-    describe("Re-INVITE with SDP failure...", () => {
-      describe("Alice invite()", () => {
-        beforeEach(() => {
-          resetSpies();
-          return inviter.invite()
-            .then(() => bob.transport.waitSent());
-        });
-
-        describe("Bob accept()", () => {
-          beforeEach(() => {
-            resetSpies();
-            invitation.delegate = undefined;
-            return invitation.accept()
-              .then(() => alice.transport.waitSent()); // ACK
-          });
-
-          it("her session state should be `established`", () => {
-            expect(inviter.state).toBe(SessionState.Established);
-          });
-
-          it("his session state should be `established`", () => {
-            expect(invitation.state).toBe(SessionState.Established);
-          });
-
-          describe("Alice invite() failure", () => {
-            beforeEach(async () => {
-              resetSpies();
-              const session: Session = inviter;
-              const sessionDescriptionHandler = session.sessionDescriptionHandler;
-              if (!sessionDescriptionHandler) {
-                throw new Error("Session description handler undefined");
-              }
-              const sdh = sessionDescriptionHandler as jasmine.SpyObj<Required<SessionDescriptionHandler>>;
-              sdh.getDescription.and.callFake(() => Promise.reject(new Error("Failed to get description.")));
-              return session.invite()
-                .catch((error) => {
-                  return;
-                });
-            });
-
-            it("her ua should send nothing", () => {
-              const spy = alice.transportSendSpy;
-              expect(spy).toHaveBeenCalledTimes(0);
-            });
-
-            it("her signaling should be stable", () => {
-              if (!inviter.dialog) {
-                fail("Session dialog undefined");
-                return;
-              }
-              expect(inviter.dialog.signalingState).toEqual(SignalingState.Stable);
-            });
-
-            it("his signaling should be stable", () => {
-              if (!invitation.dialog) {
-                fail("Session dialog undefined");
-                return;
-              }
-              expect(invitation.dialog.signalingState).toEqual(SignalingState.Stable);
-            });
-          });
-        });
-      });
-    });
-
-    describe("Re-INVITE with SDP...", () => {
-      reinviteSuite(false);
-    });
-
-    describe("Re-INVITE without SDP...", () => {
-      reinviteSuite(true);
-    });
-
-    // This group of tests is probably better covered in conjunction with testing REFER w/Replaces
-    describe("INVITE with Replaces...", () => {
-      describe("Alice invite()", () => {
-        beforeEach(() => {
-          resetSpies();
-          return inviter.invite()
-            .then(() => bob.transport.waitSent());
-        });
-
-        describe("Bob accept()", () => {
-          beforeEach(() => {
-            resetSpies();
-            invitation.delegate = undefined;
-            return invitation.accept()
-              .then(() => alice.transport.waitSent()); // ACK
-          });
-
-          it("her state should be `established`", () => {
-            expect(inviter.state).toBe(SessionState.Established);
-          });
-
-          it("his state should be `established`", () => {
-            expect(invitation.state).toBe(SessionState.Established);
-          });
-
-          describe("Carol invite() with Replaces to Alice...", () => {
-            let carol: UserFake;
-            let replacesInviter: Inviter;
-            let replacesInvitation: Invitation;
-            let replacesInvitationEmitSpy: EventEmitterEmitSpy;
-            let replacesInvitationStateSpy: EmitterSpy<SessionState>;
-
-            function resetSpies3(): void {
-              resetSpies();
-              carol.transportReceiveSpy.calls.reset();
-              carol.transportSendSpy.calls.reset();
-              if (replacesInvitationEmitSpy) { replacesInvitationEmitSpy.calls.reset(); }
-              if (replacesInvitationStateSpy) { replacesInvitationStateSpy.calls.reset(); }
-            }
-
-            beforeEach(async () => {
-              carol = makeUserFake("carol", "example.com", "Carol");
-              connectUserFake(alice, carol);
-              return carol.userAgent.start();
-            });
-
-            afterEach(async () => {
-              return carol.userAgent.stop()
-                .then(() => expect(carol.isShutdown()).toBe(true));
-            });
-
-            describe("Replacing unknown session", () => {
-              beforeEach(async () => {
-                resetSpies3();
-                alice.userAgent.delegate = {
-                  onInvite: (session) => {
-                    replacesInvitation = session;
-                    replacesInvitationEmitSpy
-                      = makeEventEmitterEmitSpy(replacesInvitation, alice.userAgent.getLogger("Alice"));
-                    replacesInvitationStateSpy
-                      = makeEmitterSpy(replacesInvitation.stateChange, alice.userAgent.getLogger("Alice"));
-                  }
-                };
-                const callId = "unknown";
-                const remoteTag = invitation.request.fromTag;
-                const localTag = invitation.request.toTag;
-                const replaces = `${callId};to-tag=${remoteTag};from-tag=${localTag}`;
-                const options = {
-                  extraHeaders: ["Replaces: " + replaces]
-                };
-                replacesInviter = new Inviter(carol.userAgent, alice.uri, options);
-                return replacesInviter.invite()
-                  .catch((error) => {
-                    return;
-                  });
-              });
-
-              it("Carol ua should send INVITE, ACK", () => {
-                const spy = carol.transportSendSpy;
-                expect(spy).toHaveBeenCalledTimes(2);
-                expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-                expect(spy.calls.argsFor(1)).toEqual(SIP_ACK);
-              });
-
-              it("Carol ua should receive 100, 481", () => {
-                const spy = carol.transportReceiveSpy;
-                expect(spy).toHaveBeenCalledTimes(2);
-                expect(spy.calls.argsFor(0)).toEqual(SIP_100);
-                expect(spy.calls.argsFor(1)).toEqual(SIP_481);
-              });
-
-              it("Carol state should be `terminated`", () => {
-                expect(replacesInviter.state).toBe(SessionState.Terminated);
-              });
-            });
-
-            describe("Replacing Bob's session", () => {
-              beforeEach(async () => {
-                resetSpies3();
-                alice.userAgent.delegate = {
-                  onInvite: (session) => {
-                    replacesInvitation = session;
-                    replacesInvitationEmitSpy =
-                      makeEventEmitterEmitSpy(replacesInvitation, alice.userAgent.getLogger("Alice"));
-                    replacesInvitationStateSpy =
-                      makeEmitterSpy(replacesInvitation.stateChange, alice.userAgent.getLogger("Alice"));
-                  }
-                };
-                const callId = invitation.request.callId;
-                const remoteTag = invitation.request.fromTag;
-                const localTag = invitation.request.toTag;
-                const replaces = `${callId};to-tag=${remoteTag};from-tag=${localTag}`;
-                const options = {
-                  extraHeaders: ["Replaces: " + replaces]
-                };
-                replacesInviter = new Inviter(carol.userAgent, alice.uri, options);
-                return replacesInviter.invite()
-                  .then(() => alice.transport.waitSent()) // provisional response
-                  .catch((error) => {
-                    return;
-                  });
-              });
-
-              it("Carol state should be `establishing`", () => {
-                expect(replacesInviter.state).toBe(SessionState.Establishing);
-              });
-
-              describe("Alice accept()", () => {
-                beforeEach(async () => {
-                  return replacesInvitation.accept()
-                    .then(() => carol.transport.waitSent()); // ACK
-                });
-
-                it("Carol ua should send INVITE, ACK", () => {
-                  const spy = carol.transportSendSpy;
-                  expect(spy).toHaveBeenCalledTimes(3);
-                  expect(spy.calls.argsFor(0)).toEqual(SIP_INVITE);
-                  expect(spy.calls.argsFor(1)).toEqual(SIP_404); // To the BYE to Bob, we get a copy
-                  expect(spy.calls.argsFor(2)).toEqual(SIP_ACK);
-                });
-
-                it("Carol ua should receive 100, 180, 200", () => {
-                  const spy = carol.transportReceiveSpy;
-                  expect(spy).toHaveBeenCalledTimes(4);
-                  expect(spy.calls.argsFor(0)).toEqual(SIP_100);
-                  expect(spy.calls.argsFor(1)).toEqual(SIP_180);
-                  expect(spy.calls.argsFor(2)).toEqual(SIP_200);
-                  expect(spy.calls.argsFor(3)).toEqual(SIP_BYE); // This is to Bob, but we get a copy
-                });
-
-                it("Carol state should be established and stable", () => {
-                  if (!replacesInviter.dialog) {
-                    fail("Session dialog undefined");
-                    return;
-                  }
-                  expect(replacesInviter.state).toBe(SessionState.Established);
-                  expect(replacesInviter.dialog.signalingState).toEqual(SignalingState.Stable);
-                });
-
-                it("Alice state to be established and stable", () => {
-                  if (!replacesInvitation.dialog) {
-                    fail("Session dialog undefined");
-                    return;
-                  }
-                  expect(replacesInvitation.state).toBe(SessionState.Established);
-                  expect(replacesInvitation.dialog.signalingState).toEqual(SignalingState.Stable);
-                });
-
-                it("Alice to Bob state should be terminated", () => {
-                  expect(inviter.state).toBe(SessionState.Terminated);
-                  expect(invitation.state).toBe(SessionState.Terminated);
-                });
-              });
-            });
-          });
-        });
-      });
     });
   });
 });
